# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "User Management API",
      "description": "API for managing user data"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "tags": [
      {
        "name": "Users",
        "description": "API for managing users"
      }
    ],
    "paths": {
      "/users": {
        "get": {
          "tags": ["Users"],
          "summary": "Get all users",
          "responses": {
            "200": {
              "description": "List of users",
              "schema": {
                "type": "array",
                "items": { "$ref": "#/definitions/User" }
              }
            }
          }
        },
        "post": {
          "tags": ["Users"],
          "summary": "Create a new user",
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "required": true,
              "schema": { "$ref": "#/definitions/UserInput" }
            }
          ],
          "responses": {
            "201": { "description": "User created successfully" },
            "400": { "description": "Validation error" }
          }
        }
      },
      "/users/{email}": {
        "get": {
          "tags": ["Users"],
          "summary": "Get a user by email",
          "parameters": [
            { "name": "email", "in": "path", "required": true, "type": "string" }
          ],
          "responses": {
            "200": { "description": "User data", "schema": { "$ref": "#/definitions/User" } },
            "404": { "description": "User not found" }
          }
        },
        "put": {
          "tags": ["Users"],
          "summary": "Update a user by email",
          "parameters": [
            { "name": "email", "in": "path", "required": true, "type": "string" },
            {
              "in": "body",
              "name": "body",
              "required": true,
              "schema": { "$ref": "#/definitions/UserInput" }
            }
          ],
          "responses": {
            "200": { "description": "User updated successfully" },
            "404": { "description": "User not found" }
          }
        },
        "delete": {
          "tags": ["Users"],
          "summary": "Delete a user by email",
          "parameters": [
            { "name": "email", "in": "path", "required": true, "type": "string" }
          ],
          "responses": {
            "200": { "description": "User deleted successfully" },
            "404": { "description": "User not found" }
          }
        }
      }
    },
    "definitions": {
      "User": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "class": { "type": "string" }
        }
      },
      "UserInput": {
        "type": "object",
        "properties": {
          "email":{"type":"string"},
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "class": { "type": "string" }
        }
      }
    }
  }